Make threshold changing working in the same way as for example color changing (regarding coding).

+++ voltex/Renderer.java
@@ -14,18 +14,18 @@ abstract public class Renderer {
 
 	protected ImagePlus image;
 	protected Volume volume;
-	protected double threshold = 0.0;
+	protected int threshold = 0;
 
 	public Renderer(ImagePlus image) {
 		this.image = image;
 		this.volume = new Volume(image);
 	}
 
-	public double getThreshold() {
+	public int  getThreshold() {
 		return threshold;
 	}
 
-	abstract public void setThreshold(double d);
+	abstract public void setThreshold(int d);
 
 	/** 
 	 * Retrieve the branchgroup of this renderer
